fix(audit): resolve all 15 E2E audit defects + supporting improvements
E2E audit (docs/E2E_AUDIT_2026_05_17.md) found 15 defects spanning
worker, Triage parser, judge LLM, mediator, LTM, deterministic layers,
and frontend. Each fix below has unit/integration coverage and was
verified end-to-end on a fresh live API run.

Backend deterministic fixes:
- CWD-01: resolve_data() helper anchors data/sigma_rules to repo root
  regardless of worker CWD (2946 Sigma rules now load from apps/api/)
- T-01: Triage parser accepts string-form sample with sha256 regex
  validation + overview.json sha256/name/size backfill
- NEG-01: enrichment exceptions promoted to WARNING, added post-
  enrichment summary line (signatures/dns/http/tcp/udp counts)
- OBS-01: worker temp file keeps original suffix (.elf VM profile)
- TIME-01: started_at/completed_at use Postgres func.now() via UPDATE
- IOC-01: IPv4 filter rejects X.509 OIDs, RFC1918/3927/5737, multicast,
  loopback (no more 2.5.4.62 in auto-generated YARA strings)
- E-01: pyelftools dependency + DT_NEEDED/.dynsym import & export
  extraction for ELF binaries
- W-01 (permanent): mock mode requires BOTH MALJAN_MOCK_MODE env AND
  settings.mock_mode_allowed (prevents silent fallback on env leak);
  every job logs Pipeline mode: REAL/MOCK with all three gates

Judge LLM hardening (new src/maljan/agents/judge_postprocess.py):
- J-01: invalid STIX IDs rewritten (placeholder UUIDs -> uuid4,
  T#### IDs -> deterministic uuid5(MITRE_NS)); source_ref/target_ref/
  object_refs updated via id_remap dict
- J-02: hallucinated indicators dropped post-LLM via evidence-corpus
  filter; system prompt forbids inferred patterns
- REP-01: attack-pattern external_references back-filled from
  MITRE_LOOKUP table for happy path (was fallback-only)
- MED-01: mediator system prompt rewritten to forbid verdict prose
  (no more "CLEAN / NO THREAT" leaking into judge context)

Memory and timing:
- LTM-01: Qdrant LTM quality gate skips low-signal runs;
  retrieve() exclude_sample_id prevents self-amplification
- A-01: react_agent_timeout_overrides config + static analyst gets
  600s; Ghidra tool allowlist trimmed 31 -> 12 high-value tools

Frontend and API:
- LIVE-01: Redis Stream xadd alongside pubsub (maxlen=1000, 24h TTL);
  new GET /jobs/{id}/events endpoint; live page backfills on mount
  with stream_id-based dedupe via useRef Set

Supporting improvements bundled in:
- Auth: PATCH /auth/me endpoint + tests
- Web: SearchPalette component, dashboard/jobs/samples/settings polish
- Migrations: audit resource_id type fix, multi-user sample dedup +
  numeric duration columns
- API: stricter upload MIME allow-list (ELF/PE synonyms, MSI, JAR, etc.)

Validation: pytest 1056 passed, mypy 0 errors, ruff clean, tsc 0 errors.
Live API run on zararli.elf: 7 TTPs (was 1), 70 STIX SDOs with 0 invalid
IDs and 0 hallucinated indicators, all attack-patterns carry populated
external_references.

+++ apps/api/app/models/job.py
@@ -3,7 +3,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, Text
+from sqlalchemy import DateTime, ForeignKey, Numeric, String, Text
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -31,7 +31,10 @@ class AnalysisJob(UUIDPrimaryKeyMixin, TimestampMixin, Base):
     # Timestamps
     started_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
     completed_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
-    duration_seconds: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    # ``Numeric(10, 2)`` so sub-second analyses (e.g. mock pipeline at ~0.7s)
+    # survive serialization. Integer truncated to 0 and broke the dashboard
+    # "AVG DURATION" tile + analysis page duration badge.
+    duration_seconds: Mapped[float | None] = mapped_column(Numeric(10, 2), nullable=True)
 
     # Configuration snapshot (LLM provider, max_iterations, etc.)
     config: Mapped[dict | None] = mapped_column(JSONB, nullable=True)
